requirements.txt
langchain
langchain-community
openai 
arxiv 
faiss-cpu 
pinecone-client 
streamlit
tiktoken